requirements.txt
# Core ML/DL Libraries
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.1

# Medical Imaging
nibabel==5.1.0
SimpleITK==2.2.1
pydicom==2.4.2

# Image Processing
opencv-python==4.8.0.74
Pillow==10.0.0
albumentations==1.3.1

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# Hyperparameter Optimization
optuna==3.3.0

# Progress Bars
tqdm==4.65.0

# Web Framework (Backend)
flask==2.3.3
flask-cors==4.0.0
werkzeug==2.3.7

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.0

# Statistical Analysis
statsmodels==0.14.0

# Optional: GPU Support
# torch==2.0.1+cu118 (if using CUDA)
# torchvision==0.15.2+cu118 (if using CUDA)
